Allow creating a py EagerTensor that shares the underlying TensorHandle.
This is so that gradients with respect to scalars pass (see the test added in
backprop_test.py).

A micro benchmark just calling constant_op.constant slows down a bit - this is
inevitable as we are creating a new python object.
After: walltime: ~2.1
Before: walltime: ~1.47

Linear regression benchmark is pretty much unchanged.

PiperOrigin-RevId: 211753801