Don't use tensorflow::Edge after freeing it
Even with this bug we were accidentally doing the right thing (so the test case
doesn't actually fail without the fix): deleting an Edge sets its input and
output indices to kControlSlot-1 so we'd normally expect to fail when there is a
control edge out of the TF cluster (because a control edge would be recognized
as a data edge).  But AddEdge(x, -1, y, -1) seems to do the right thing for both
control and data edges.

PiperOrigin-RevId: 214831204